+++ src/RecoveryServices/RecoveryServices.Backup.ServiceClientAdapter/BMSAPIs/RecoveryPointsAPIs.cs
@@ -1,4 +1,4 @@
-﻿// ----------------------------------------------------------------------------------
+// ----------------------------------------------------------------------------------
 //
 // Copyright Microsoft Corporation
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -42,12 +42,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string resourceGroupName = null)
         {
             var response = BmsAdapter.Client.RecoveryPoints.GetWithHttpMessagesAsync(
-                vaultName ?? BmsAdapter.GetResourceName(),
-                resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                AzureFabricName,
-                containerName,
-                protectedItemName,
-                recoveryPointId,
+                vaultName: vaultName ?? BmsAdapter.GetResourceName(),
+                resourceGroupName: resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                fabricName: AzureFabricName,
+                containerName: containerName,
+                protectedItemName: protectedItemName,
+                recoveryPointId: recoveryPointId,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             return response;
@@ -71,17 +71,17 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
         {
             Func<RestAzureNS.IPage<RecoveryPointResource>> listAsync =
                 () => BmsAdapter.Client.RecoveryPoints.ListWithHttpMessagesAsync(
-                    vaultName ?? BmsAdapter.GetResourceName(),
-                    resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                    AzureFabricName,
-                    containerName,
-                    protectedItemName,
-                    queryFilter,
+                    vaultName: vaultName ?? BmsAdapter.GetResourceName(),
+                    resourceGroupName: resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                    fabricName: AzureFabricName,
+                    containerName: containerName,
+                    protectedItemName: protectedItemName,
+                    filter: queryFilter?.Filter,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             Func<string, RestAzureNS.IPage<RecoveryPointResource>> listNextAsync =
                 nextLink => BmsAdapter.Client.RecoveryPoints.ListNextWithHttpMessagesAsync(
-                    nextLink,
+                    nextPageLink: nextLink,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             var response = HelperUtils.GetPagedList(listAsync, listNextAsync);
@@ -106,17 +106,17 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
         {
             Func<RestAzureNS.IPage<CrrModel.RecoveryPointResource>> listAsync =
                 () => CrrAdapter.Client.RecoveryPointsCrr.ListWithHttpMessagesAsync(
-                    vaultName ?? BmsAdapter.GetResourceName(),
-                    resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                    AzureFabricName,
-                    containerName,
-                    protectedItemName,
-                    queryFilter,
+                    vaultName: vaultName ?? BmsAdapter.GetResourceName(),
+                    resourceGroupName: resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                    fabricName: AzureFabricName,
+                    containerName: containerName,
+                    protectedItemName: protectedItemName,
+                    odataQuery: queryFilter,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             Func<string, RestAzureNS.IPage<CrrModel.RecoveryPointResource>> listNextAsync =
                 nextLink => CrrAdapter.Client.RecoveryPointsCrr.ListNextWithHttpMessagesAsync(
-                    nextLink,
+                    nextPageLink: nextLink,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             var response = HelperUtils.GetPagedListCrr(listAsync, listNextAsync);
@@ -140,12 +140,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string resourceGroupName = null)
         {
             var response = CrrAdapter.Client.RecoveryPointsCrr.GetWithHttpMessagesAsync(
-                vaultName ?? BmsAdapter.GetResourceName(),
-                resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                AzureFabricName,
-                containerName,
-                protectedItemName,
-                recoveryPointId,
+                vaultName: vaultName ?? BmsAdapter.GetResourceName(),
+                resourceGroupName: resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                fabricName: AzureFabricName,
+                containerName: containerName,
+                protectedItemName: protectedItemName,
+                recoveryPointId: recoveryPointId,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             return response;
@@ -169,18 +169,17 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
         {
             Func<RestAzureNS.IPage<RecoveryPointResource>> listAsync =
                 () => BmsAdapter.Client.RecoveryPointsRecommendedForMove.ListWithHttpMessagesAsync(
-                vaultName,
-                resourceGroupName,
-                AzureFabricName,
-                containerName,
-                protectedItemName,
-                moveRequest.ObjectType,
-                moveRequest.ExcludedRPList
-                ).Result.Body;
+                    vaultName: vaultName,
+                    resourceGroupName: resourceGroupName,
+                    fabricName: AzureFabricName,
+                    containerName: containerName,
+                    protectedItemName: protectedItemName,
+                    objectType: moveRequest.ObjectType,
+                    excludedRPList: moveRequest.ExcludedRPList).Result.Body;
 
             Func<string, RestAzureNS.IPage<RecoveryPointResource>> listNextAsync =
                 nextLink => BmsAdapter.Client.RecoveryPointsRecommendedForMove.ListNextWithHttpMessagesAsync(
-                    nextLink,
+                    nextPageLink: nextLink,
                     cancellationToken: BmsAdapter.CmdletCancellationToken).Result.Body;
 
             var response = HelperUtils.GetPagedList(listAsync, listNextAsync);
@@ -205,15 +204,20 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string vaultName = null,
             string resourceGroupName = null)
         {
-            return BmsAdapter.Client.BeginMoveRecoveryPointWithHttpMessagesAsync(
-                vaultName,
-                resourceGroupName,
-                AzureFabricName,
-                containerName,
-                protectedItemName,
-                recoveryPointId,
-                moveRPAcrossTiersRequest
-                ).Result;
+            var response = BmsAdapter.Client.BeginMoveRecoveryPointWithHttpMessagesAsync(
+                vaultName: vaultName,
+                resourceGroupName: resourceGroupName,
+                fabricName: AzureFabricName,
+                containerName: containerName,
+                protectedItemName: protectedItemName,
+                recoveryPointId: recoveryPointId,
+                parameters: moveRPAcrossTiersRequest).Result;
+            return new RestAzureNS.AzureOperationResponse
+            {
+                Request = response.Request,
+                Response = response.Response,
+                RequestId = response.RequestId
+            };
         }
 
         /// <summary>
@@ -238,13 +242,13 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             provisionRequest.Properties = registrationRequest;
 
             var response = BmsAdapter.Client.ItemLevelRecoveryConnections.ProvisionWithHttpMessagesAsync(
-                vaultName ?? BmsAdapter.GetResourceName(),
-                resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                AzureFabricName,
-                containerName,
-                protectedItemName,
-                recoveryPointId,
-                provisionRequest,
+                vaultName: vaultName ?? BmsAdapter.GetResourceName(),
+                resourceGroupName: resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                fabricName: AzureFabricName,
+                containerName: containerName,
+                protectedItemName: protectedItemName,
+                recoveryPointId: recoveryPointId,
+                parameters: provisionRequest,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result;
 
             return response;
@@ -267,12 +271,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Cmdlets.ServiceClient
             string resourceGroupName = null)
         {
             var response = BmsAdapter.Client.ItemLevelRecoveryConnections.RevokeWithHttpMessagesAsync(
-                vaultName ?? BmsAdapter.GetResourceName(),
-                resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
-                AzureFabricName,
-                containerName,
-                protectedItemName,
-                recoveryPointId,
+                vaultName: vaultName ?? BmsAdapter.GetResourceName(),
+                resourceGroupName: resourceGroupName ?? BmsAdapter.GetResourceGroupName(),
+                fabricName: AzureFabricName,
+                containerName: containerName,
+                protectedItemName: protectedItemName,
+                recoveryPointId: recoveryPointId,
                 cancellationToken: BmsAdapter.CmdletCancellationToken).Result;
 
             return response;